+++ addon/components/select-box/index.js
@@ -41,7 +41,7 @@ import {
   registerOptionsContainer
 } from '../../utils/registration/options-container';
 import { focusOut } from '../../utils/select-box/focus';
-import { keyDown, keyPress, pressEnter } from '../../utils/select-box/keyboard';
+import { keyUp, keyPress, pressEnter } from '../../utils/select-box/keyboard';
 import { setInputValue } from '../../utils/select-box/input/value';
 import buildAPI from '../../utils/shared/api';
 import {
@@ -276,8 +276,8 @@ export default class SelectBox extends Component {
   }
 
   @action
-  handleKeyDown(e) {
-    keyDown(this, e);
+  handleKeyUp(e) {
+    keyUp(this, e);
   }
 
   @action